refactor: Structure of project
