+++ stcli/__init__.py
@@ -1,3 +1,3 @@
 from stcli import config
-from stcli import lib
 from stcli import src
+from stcli.src import lib
