+++ stcli/src/lib/cli.py
@@ -1,7 +1,7 @@
 import click
 import colorama
 
-from stcli.lib.error import print_output_error
+from stcli.src.lib.error import print_output_error
 
 from stcli.src import info
 from stcli.src import generator as creator
