+++ stcli/src/lib/error.py
